requirements.txt
AccessControl==7.2
Acquisition==6.1
aiofiles==24.1.0
alembic==1.17.2
altgraph==0.17.4
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.7.0
AnyQt==0.2.0
APScheduler==3.11.2
asgiref==3.8.1
attrs==25.1.0
AuthEncoding==5.0
autocommand==2.2.2
Automat==24.8.1
babel==2.17.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.42.21
botocore==1.42.21
bottle==0.13.2
BTrees==6.1
cachetools==5.5.0
Cerberus==1.3.5
certifi==2024.12.14
cffi==1.17.1
Chameleon==4.5.4
chardet==5.2.0
charset-normalizer==3.4.0
cheroot==10.0.1
CherryPy==18.10.0
circuits==3.2.3
click==8.1.8
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.1
cssselect==1.2.0
CTkMessagebox==2.7
cubicweb==4.8.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.5
dectate==0.14
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.4
dnspython==2.7.0
docker==7.1.0
DocumentTemplate==4.6
docutils==0.21.2
Eve==2.2.0
Events==0.3
ExtensionClass==6.0
falcon==2.0.0
fastapi==0.115.6
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
fsspec==2025.2.0
greenlet==3.3.0
grok==5.1
grokcore.annotation==4.0
grokcore.catalog==4.0
grokcore.chameleon==4.1
grokcore.component==4.1
grokcore.content==4.1
grokcore.formlib==4.0
grokcore.layout==4.0
grokcore.message==4.0
grokcore.security==4.0
grokcore.site==4.0
grokcore.traverser==4.0
grokcore.view==4.0
grokcore.viewlet==4.0
growler==0.8.0
h11==0.14.0
html2text==2024.2.26
html5tagger==1.3.0
httptools==0.6.4
hug==2.6.1
hupper==1.12.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
importscan==0.2
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.5.1
jsonpickle==4.1.1
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.7
legacy-cgi==2.6.1
logilab-common==1.11.0
logilab-constraint==0.8.0
logilab-database==1.20.0
logilab-mtconverter==0.12.0
lxml==5.3.0
lxml_html_clean==0.4.1
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
martian==2.0.post1
matplotlib==3.10.0
mdurl==0.1.2
MIDIUtil==1.2.1
more-itertools==10.5.0
morepath==0.19
mpmath==1.3.0
multidict==6.1.0
MultiMapping==5.0
multipart==1.2.1
music21==9.7.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
optional-django==0.3.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsel==1.10.0
passlib==1.7.4
PasteDeploy==3.1.0
patsy==1.0.1
pefile==2023.2.7
Persistence==5.1
persistent==6.1
pillow==11.0.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.3.8
plotly==5.24.1
portend==3.2.0
Protego==0.4.0
psycopg-binary==3.2.13
psycopg2==2.9.10
psycopg2-binary==2.9.11
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycnic==0.1.9
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyDispatcher==2.0.7
PyForms==3.0.0
pygame==2.6.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
pylance==0.23.1
pymongo==4.10.1
pynput==1.7.7
PyOpenGL==3.1.7
pyOpenSSL==25.0.0
pyparsing==3.2.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.2
pyramid==1.10.8
pyramid-multiauth==0.9.0
PySimpleGUI==5.0.7
PySocks==1.7.1
PySoundFile==0.9.0.post1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-gettext==5.0
python-multipart==0.0.20
python-rtmidi==1.5.8
pytz==2024.2
pywin32==310
pywin32-ctypes==0.2.3
pyzero==0.1
QScintilla==2.14.1
queuelib==1.7.0
Quixote==3.6
rdflib==7.1.1
react==4.3.0
reg==0.12
repoze.lru==0.7
requests==2.32.3
requests-file==2.1.0
requirements.py==1.0.1
RestrictedPython==7.4
retrying==1.3.4
rich==13.9.4
roman==4.2
rql==0.43.2
rsa==4.9
s3transfer==0.16.0
sanic==24.6.0
sanic-routing==23.12.0
scipy==1.15.2
Scrapy==2.12.0
seaborn==0.13.2
selenium==4.29.0
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.45
sqlparse==0.5.3
starlette==0.41.3
statsmodels==0.14.4
sympy==1.13.1
tempora==5.7.0
tenacity==9.0.0
tkcalendar==1.6.1
tldextract==5.1.3
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tracerite==1.1.1
transaction==5.0
translationstring==1.4
trio==0.29.0
trio-websocket==0.12.1
Twisted==24.11.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
venusian==3.1.1
virtualenv==20.31.2
visvis==1.14.0
w3lib==2.3.1
waitress==3.0.2
webcolors==24.11.1
WebOb==1.8.9
websocket-client==1.8.0
websockets==14.1
WebTest==3.0.2
Werkzeug==3.0.6
wheel==0.45.1
wsgicors==0.7.0
WSGIProxy2==0.5.1
wsproto==1.2.0
wxPython==4.2.2
yams==0.54.0
Yapps2-logilab==2.2.1
z3c.autoinclude==1.0
z3c.flashmessage==3.0
z3c.pt==4.4
zc.catalog==3.0
zc.lockfile==3.0.post1
ZConfig==4.1
zdaemon==5.1
zExceptions==5.0
zipp==3.21.0
ZODB==6.0
zodbpickle==4.1.1
Zope==5.11.1
zope.annotation==5.0
zope.app.appsetup==5.0
zope.app.publication==5.0
zope.app.wsgi==5.3
zope.authentication==5.0
zope.browser==3.0
zope.browsermenu==5.0
zope.browserpage==5.0
zope.browserresource==5.1
zope.cachedescriptors==5.0
zope.catalog==5.0
zope.component==6.0
zope.configuration==6.0
zope.container==6.1
zope.contentprovider==6.0
zope.contenttype==5.1
zope.datetime==5.0.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.dottedname==6.0
zope.error==5.0
zope.errorview==2.0
zope.event==5.0
zope.exceptions==5.2
zope.filerepresentation==6.0
zope.formlib==6.1
zope.generations==5.1.0
zope.globalrequest==2.0
zope.hookable==7.0
zope.i18n==5.2
zope.i18nmessageid==7.0
zope.index==7.0
zope.interface==7.2
zope.intid==5.0
zope.keyreference==6.0
zope.lifecycleevent==5.0
zope.location==5.0
zope.login==3.0
zope.minmax==2.3
zope.pagetemplate==5.1
zope.password==4.4
zope.principalregistry==5.0
zope.processlifetime==3.0
zope.proxy==6.1
zope.ptresource==5.0
zope.publisher==7.2
zope.schema==7.0.1
zope.security==7.3
zope.securitypolicy==5.0
zope.sequencesort==5.0
zope.session==5.1
zope.site==5.0
zope.size==5.0
zope.structuredtext==5.0
zope.tal==5.0.1
zope.tales==6.0
zope.testbrowser==7.0
zope.testing==5.0.1
zope.traversing==5.0
zope.viewlet==5.0
